private shrink and serialize functions
