+++ qiniu/httpc/region.js
@@ -167,114 +167,6 @@ Region.fromRegionId = function (regionId, options) {
     });
 };
 
-/**
- * @typedef RegionPersistInfo
- * @property {string} [regionId]
- * @property {string} s3RegionId
- * @property {Object.<string, EndpointPersistInfo[]>} services
- * @property {number} ttl
- * @property {number} createTime
- */
-
-/**
- * @param {RegionPersistInfo} persistInfo
- * @returns {Region}
- */
-Region.fromPersistInfo = function (persistInfo) {
-    /**
-     * @param {EndpointPersistInfo[]} servicePersistEndpoint
-     * @returns {Endpoint[]}
-     */
-    const convertToEndpoints = (servicePersistEndpoint) => {
-        // The `persistInfo` is from disk that may be broken.
-        if (!Array.isArray(servicePersistEndpoint)) {
-            return [];
-        }
-
-        return servicePersistEndpoint.map(e => Endpoint.fromPersistInfo(e));
-    };
-
-    /**
-     * @type {Object.<ServiceKey, Endpoint[]>}
-     */
-    const services = {};
-    for (const serviceName of Object.keys(persistInfo.services)) {
-        const endpointPersistInfos = persistInfo.services[serviceName];
-        services[serviceName] = convertToEndpoints(endpointPersistInfos);
-    }
-
-    return new Region({
-        regionId: persistInfo.regionId,
-        s3RegionId: persistInfo.s3RegionId,
-        services: services,
-        ttl: persistInfo.ttl,
-        createTime: new Date(persistInfo.createTime)
-    });
-};
-
-/**
- * @param {Object} data
- * @param {string} data.region
- * @param {Object} data.s3
- * @param {string[]} data.s3.domains
- * @param {string} data.s3.region_alias
- * @param {Object} data.uc
- * @param {string[]} data.uc.domains
- * @param {Object} data.up
- * @param {string[]} data.up.domains
- * @param {Object} data.io
- * @param {string[]} data.io.domains
- * @param {Object} data.rs
- * @param {string[]} data.rs.domains
- * @param {Object} data.rsf
- * @param {string[]} data.rsf.domains
- * @param {Object} data.api
- * @param {string[]} data.api.domains
- * @param {number} data.ttl
- * @returns {Region}
- */
-Region.fromQueryData = function (data) {
-    /**
-     * @param {string[]} domains
-     * @returns {Endpoint[]}
-     */
-    const convertToEndpoints = (domains) => {
-        if (!Array.isArray(domains)) {
-            return [];
-        }
-        return domains.map(d => new Endpoint(d));
-    };
-
-    let services = {
-        [SERVICE_NAME.UC]: convertToEndpoints(data.uc.domains),
-        [SERVICE_NAME.UP]: convertToEndpoints(data.up.domains),
-        [SERVICE_NAME.IO]: convertToEndpoints(data.io.domains),
-        [SERVICE_NAME.RS]: convertToEndpoints(data.rs.domains),
-        [SERVICE_NAME.RSF]: convertToEndpoints(data.rsf.domains),
-        [SERVICE_NAME.API]: convertToEndpoints(data.api.domains),
-        [SERVICE_NAME.S3]: convertToEndpoints(data.s3.domains)
-    };
-
-    // forward compatibility with new services
-    services = Object.keys(data)
-        // use Object.entries when min version of Node.js update to ≥ v7.5.0
-        .map(k => ([k, data[k]]))
-        .reduce((s, [k, v]) => {
-            if (v && Array.isArray(v.domains) && !(k in s)) {
-                s[k] = convertToEndpoints(v.domains);
-            }
-            return s;
-        }, services);
-
-    return new Region({
-        regionId: data.region,
-        s3RegionId: data.s3.region_alias,
-        services: services,
-        ttl: data.ttl,
-        createTime: new Date()
-    });
-};
-
 Object.defineProperty(Region.prototype, 'isLive', {
     get: function () {
         if (this.ttl < 0) {
@@ -283,33 +175,6 @@ Object.defineProperty(Region.prototype, 'isLive', {
         // convert ms to s
         const liveTime = Math.round((Date.now() - this.createTime) / 1000);
         return liveTime < this.ttl;
-    },
-    enumerable: false,
-    configurable: true
-});
-
-Object.defineProperty(Region.prototype, 'persistInfo', {
-    /**
-     * @returns {RegionPersistInfo}
-     */
-    get: function () {
-        /**
-         * @type {Object.<string, EndpointPersistInfo[]>}
-         */
-        const persistedServices = {};
-        // use Object.entries when min version of Node.js update to ≥ v7.5.0
-        for (const k of Object.keys(this.services)) {
-            const v = this.services[k];
-            persistedServices[k] = v.map(endpoint => endpoint.persistInfo);
-        }
-
-        return {
-            regionId: this.regionId,
-            s3RegionId: this.s3RegionId,
-            services: persistedServices,
-            ttl: this.ttl,
-            createTime: this.createTime.getTime()
-        };
     },
     enumerable: false,
     configurable: true
