+++ test/httpc.test.js
@@ -245,27 +245,6 @@ describe('test http module', function () {
             should.equal(endpoint.getValue(), 'http://www.qiniu.com');
             should.equal(endpoint.getValue({ scheme: 'https' }), 'https://www.qiniu.com');
         });
-
-        it('test persistInfo', function () {
-            const endpoint = new Endpoint('www.qiniu.com');
-            const expectedResult = {
-                host: 'www.qiniu.com',
-                defaultScheme: 'https'
-            };
-            should.deepEqual(endpoint.persistInfo, expectedResult);
-        });
-
-        it('test fromPersistedInfo', function () {
-            const persistedData = {
-                host: 'www.qiniu.com',
-                defaultScheme: 'https'
-            };
-
-            const endpoint = Endpoint.fromPersistInfo(persistedData);
-            should.equal(endpoint.host, 'www.qiniu.com');
-            should.equal(endpoint.defaultScheme, 'https');
-            should.equal(endpoint.getValue(), 'https://www.qiniu.com');
-        });
     });
 
     describe('test region', function () {
@@ -458,133 +437,6 @@ describe('test http module', function () {
             should.equal(regionZ1.ttl, -1);
             should.equal(regionZ1.createTime.getTime(), 0);
         });
-
-        it('test persistInfo', function () {
-            const now = new Date();
-            const regionZ0 = Region.fromRegionId(
-                'z0',
-                {
-                    createTime: now
-                }
-            );
-
-            const expectedResult = {
-                regionId: regionZ0.regionId,
-                s3RegionId: regionZ0.s3RegionId,
-                services: Object.keys(regionZ0.services)
-                    .reduce((services, serviceKey) => {
-                        const service = regionZ0.services[serviceKey];
-                        services[serviceKey] = service.map(s => ({
-                            defaultScheme: s.defaultScheme,
-                            host: s.host
-                        }));
-                        return services;
-                    }, {}),
-                ttl: 86400,
-                createTime: now.getTime()
-            };
-            should.deepEqual(regionZ0.persistInfo, expectedResult);
-        });
-
-        it('test fromPersistInfo', function () {
-            const now = new Date();
-
-            const persistInfo = {
-                regionId: 'z0',
-                s3RegionId: 'z0',
-                services: {
-                    uc: [
-                        {
-                            defaultScheme: 'https',
-                            host: 'uc.qiniuapi.com'
-                        }
-                    ],
-                    up: [
-                        {
-                            defaultScheme: 'https',
-                            host: 'upload.qiniup.com'
-                        },
-                        {
-                            defaultScheme: 'https',
-                            host: 'up.qiniup.com'
-                        },
-                        {
-                            defaultScheme: 'https',
-                            host: 'up.qbox.me'
-                        }
-                    ],
-                    io: [
-                        {
-                            defaultScheme: 'https',
-                            host: 'iovip.qiniuio.com'
-                        },
-                        {
-                            defaultScheme: 'https',
-                            host: 'iovip.qbox.me'
-                        }
-                    ],
-                    rs: [
-                        {
-                            defaultScheme: 'https',
-                            host: 'rs-z0.qiniuapi.com'
-                        },
-                        {
-                            defaultScheme: 'https',
-                            host: 'rs-z0.qbox.me'
-                        }
-                    ],
-                    rsf: [
-                        {
-                            defaultScheme: 'https',
-                            host: 'rsf-z0.qiniuapi.com'
-                        },
-                        {
-                            defaultScheme: 'https',
-                            host: 'rsf-z0.qbox.me'
-                        }
-                    ],
-                    api: [
-                        {
-                            defaultScheme: 'https',
-                            host: 'api-z0.qiniuapi.com'
-                        },
-                        {
-                            defaultScheme: 'https',
-                            host: 'api-z0.qbox.me'
-                        }
-                    ],
-                    s3: [
-                        {
-                            defaultScheme: 'https',
-                            host: 's3.z0.qiniucs.com'
-                        }
-                    ]
-                },
-                ttl: 86400,
-                createTime: now.getTime()
-            };
-
-            const region = Region.fromPersistInfo(persistInfo);
-
-            should.equal(region.regionId, 'z0');
-            should.equal(region.s3RegionId, 'z0');
-            should.equal(region.ttl, 86400);
-            should.equal(region.createTime.getTime(), now.getTime());
-
-            // use Object.entries when min version of Node.js update to ≥ v7.5.0
-            for (const serviceName of Object.keys(persistInfo.services)) {
-                const persistEndpoints = persistInfo.services[serviceName];
-                const endpointValues = region.services[serviceName].map(e => ({
-                    defaultScheme: e.defaultScheme,
-                    host: e.host
-                }));
-                const expectedEndpointValues = persistEndpoints.map(e => ({
-                    defaultScheme: e.defaultScheme,
-                    host: e.host
-                }));
-                should.deepEqual(endpointValues, expectedEndpointValues);
-            }
-        });
     });
 
     describe('test endpoints provider', function () {
@@ -690,11 +542,15 @@ describe('test http module', function () {
                 ])
                     .then(() => {
                         const content = fs.readFileSync(cachedRegionsProvider.persistPath);
+                        const actual = JSON.parse(content.toString());
                         should.deepEqual(
-                            JSON.parse(content.toString()),
+                            {
+                                cacheKey: actual.cacheKey,
+                                regions: actual.regions.map(r => r.regionId)
+                            },
                             {
                                 cacheKey,
-                                regions: [rZ0, rZ1].map(r => r.persistInfo)
+                                regions: [rZ0, rZ1].map(r => r.regionId)
                             }
                         );
                         return cachedRegionsProvider.getRegions();
@@ -748,204 +604,6 @@ describe('test http module', function () {
                     })
                     .then(regions => {
                         should.deepEqual(regions.map(r => r.regionId), ['z0', 'z1']);
-                    });
-            });
-
-            it('test CachedRegionsProvider shrink', function () {
-                const cachedRegionsProvider = getCachedRegionsProvider();
-
-                const rZ2 = Region.fromRegionId('z2');
-                const rZ3 = Region.fromRegionId('z3');
-                rZ2.createTime = new Date(0);
-                rZ3.createTime = new Date(0);
-
-                const rZ0 = Region.fromRegionId('z0');
-                const rZ1 = Region.fromRegionId('z1');
-                rZ0.createTime = new Date(0);
-                should.ok(!rZ0.isLive);
-
-                return cachedRegionsProvider.setRegions([
-                    rZ2,
-                    rZ3
-                ])
-                    .then(() => {
-                        return cachedRegionsProvider.setRegions([
-                            rZ0,
-                            rZ1
-                        ]);
-                    })
-                    .then(() => {
-                        return cachedRegionsProvider.shrink();
-                    })
-                    .then(isShrunk => {
-                        should.ok(isShrunk);
-                        const cachedRegions = cachedRegionsProvider._memoCache.get(cachedRegionsProvider.cacheKey);
-                        should.deepEqual(cachedRegions.map(r => r.regionId), ['z1']);
-
-                        const content = fs.readFileSync(cachedRegionsProvider.persistPath);
-                        const jsonl = content.toString().split(os.EOL).filter(l => l.length > 0);
-                        should.equal(jsonl.length, 1);
-
-                        const [persistedCache] = jsonl;
-                        const { cacheKey, regions } = JSON.parse(persistedCache);
-                        should.equal(cacheKey, cachedRegionsProvider.cacheKey);
-                        should.equal(regions.length, 1);
-                        should.equal(regions[0].regionId, 'z1');
-                    });
-            });
-
-            it('test CachedRegionsProvider shrunk just now', function () {
-                const cachedRegionsProvider = getCachedRegionsProvider();
-
-                const rZ0 = Region.fromRegionId('z0');
-                const rZ1 = Region.fromRegionId('z1');
-                rZ0.createTime = new Date(0);
-                should.ok(!rZ0.isLive);
-                return cachedRegionsProvider.setRegions([
-                    rZ0,
-                    rZ1
-                ])
-                    .then(() => {
-                        return cachedRegionsProvider.shrink();
-                    })
-                    .then(isShrunk => {
-                        should.ok(isShrunk);
-
-                        return cachedRegionsProvider.shrink();
-                    })
-                    .then(isShrunk => {
-                        should.ok(!isShrunk);
-
-                        const cachedRegions = cachedRegionsProvider._memoCache.get(cachedRegionsProvider.cacheKey);
-                        should.deepEqual(cachedRegions.map(r => r.regionId), ['z1']);
-
-                        const content = fs.readFileSync(cachedRegionsProvider.persistPath);
-                        const jsonl = content.toString().split(os.EOL).filter(l => l.length > 0);
-                        should.equal(jsonl.length, 1);
-                    });
-            });
-
-            it('test CachedRegionsProvider shrink force', function () {
-                const cachedRegionsProvider = getCachedRegionsProvider();
-
-                const rZ0 = Region.fromRegionId('z0');
-                const rZ1 = Region.fromRegionId('z1');
-                rZ0.createTime = new Date(0);
-                should.ok(!rZ0.isLive);
-
-                return cachedRegionsProvider.setRegions([
-                    rZ0,
-                    rZ1
-                ])
-                    .then(() => {
-                        return cachedRegionsProvider.shrink();
-                    })
-                    .then(isShrunk => {
-                        should.ok(isShrunk);
-
-                        return cachedRegionsProvider.setRegions([
-                            rZ0,
-                            rZ1
-                        ]);
-                    })
-                    .then(() => {
-                        return cachedRegionsProvider.shrink();
-                    })
-                    .then(isShrunk => {
-                        should.ok(!isShrunk);
-
-                        const cachedRegions = cachedRegionsProvider._memoCache.get(cachedRegionsProvider.cacheKey);
-                        should.deepEqual(cachedRegions.map(r => r.regionId), ['z0', 'z1']);
-
-                        const content = fs.readFileSync(cachedRegionsProvider.persistPath);
-                        const jsonl = content.toString().split(os.EOL).filter(l => l.length > 0);
-                        should.equal(jsonl.length, 2);
-
-                        return cachedRegionsProvider.shrink(true);
-                    })
-                    .then(isShrunk => {
-                        should.ok(isShrunk);
-                        const cachedRegions = cachedRegionsProvider._memoCache.get(cachedRegionsProvider.cacheKey);
-                        should.deepEqual(cachedRegions.map(r => r.regionId), ['z1']);
-
-                        const content = fs.readFileSync(cachedRegionsProvider.persistPath);
-                        const jsonl = content.toString().split(os.EOL).filter(l => l.length > 0);
-                        should.equal(jsonl.length, 1);
-                    });
-            });
-
-            it('test CachedRegionsProvider shrinking', function () {
-                const cachedRegionsProvider = getCachedRegionsProvider();
-
-                const lockFilePath = cachedRegionsProvider.persistPath + '.shrink.lock';
-                const fd = fs.openSync(lockFilePath, 'w');
-                fs.closeSync(fd);
-
-                const rZ0 = Region.fromRegionId('z0');
-                const rZ1 = Region.fromRegionId('z1');
-                rZ0.createTime = new Date(0);
-                should.ok(!rZ0.isLive);
-
-                return cachedRegionsProvider.setRegions([
-                    rZ0,
-                    rZ1
-                ])
-                    .then(() => {
-                        return cachedRegionsProvider.shrink();
-                    })
-                    .then(isShrunk => {
-                        should.ok(!isShrunk);
-                        fs.unlinkSync(lockFilePath);
-                    })
-                    .catch(err => {
-                        // use finally when min version of Node.js update to ≥ v10.3.0
-                        fs.unlinkSync(lockFilePath);
-                        return Promise.reject(err);
-                    });
-            });
-
-            it('test CachedRegionsProvider merge shrunk', function () {
-                const cachedRegionsProvider = getCachedRegionsProvider();
-
-                const rZ0 = Region.fromRegionId('z0');
-                const rZ1 = Region.fromRegionId('z1');
-                const mockDate = new Date();
-                mockDate.setMinutes(1, 2, 3);
-                rZ1.createTime = mockDate;
-                const rZ1Old = Region.fromRegionId('z1');
-                rZ1Old.createTime = new Date(0);
-                const rZ2 = Region.fromRegionId('z2');
-
-                fs.appendFileSync(
-                    cachedRegionsProvider.persistPath,
-                    JSON.stringify({
-                        cacheKey: cachedRegionsProvider.cacheKey,
-                        regions: [rZ0, rZ1].map(r => r.persistInfo)
-                    }) + os.EOL
-                );
-                fs.appendFileSync(
-                    cachedRegionsProvider.persistPath,
-                    JSON.stringify({
-                        cacheKey: cachedRegionsProvider.cacheKey,
-                        regions: [rZ1Old, rZ2].map(r => r.persistInfo)
-                    }) + os.EOL
-                );
-
-                return cachedRegionsProvider.shrink()
-                    .then(isShrunk => {
-                        should.ok(isShrunk);
-                        const content = fs.readFileSync(cachedRegionsProvider.persistPath);
-                        const jsonl = content.toString().split(os.EOL).filter(l => l.length > 0);
-                        should.equal(jsonl.length, 1);
-
-                        const [persistedCache] = jsonl;
-                        const { cacheKey, regions } = JSON.parse(persistedCache);
-                        should.equal(cacheKey, cachedRegionsProvider.cacheKey);
-                        should.equal(regions.length, 3);
-                        regions.every(r => ['z0', 'z1', 'z2'].includes(r.regionId));
-                        const actualRZ1 = regions.find(r => r.regionId === 'z1');
-                        should.exists(actualRZ1);
-                        should.equal(actualRZ1.createTime, mockDate.getTime());
                     });
             });
